GenAI/requirements.txt
python-dotenv
langchain_community
fastapi
uvicorn
sse_starlette
bs4
pypdf
Chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
ipykernel
flask
langchain-huggingface
huggingface-hub
pymupdf
langgraph
pydantic>=2.0,<2.7
langchain-core>=0.1,<0.2
langchain>=0.1,<0.2

